Attendees and Schedule
We've updated the list of probable attendees. I have also put together a tentative schedule for each day—of course, the session blocks can be broken up more concretely once we determine the topics for discussion.

The downstairs rooms listed in the schedules have already been reserved through Sebastian.

For our planning purposes (i.e. making a dinner reservation), it looks like the maximum number of participants will be around 22. Assuming that not everyone would be able to make it for dinner, we could look to place a reservation for around 18 people—better to overestimate than to underestimate. Thoughts are always welcome.