Create Information Sheet
I asked Sabine for a list of hotels that we typically recommend. She gave me a list, and I thought it would be a good idea to keep this here. We could reuse this for later events (or simply share it with future visitors).

Just a draft with some information about:
- where Haus 70 is located
- connections to/from the airport
- recommended hotels

We can add to this or modify it as we see fit over time.